fix(ng): refuse connection IDs that do not fit the length byte
A QUIC long header writes the connection ID length into a single byte and the
code wrote it unchecked. A 300-byte id declared itself as 44 -- 300 & 0xff --
and the packet went out describing a length it did not have. Demonstrated by
running the same program against the old and new library: 300 bytes in, "header
declares dcid len 44" out. RFC 9000 17.2 caps an id at 20 bytes, so that is the
boundary held, for both the destination and source id.

Unreachable from this tool, which uses 8-byte ids, but initial_packet() is a
public function whose ids are parameters.

Two things checked and found not to be bugs, kept as hardening with the
comments rewritten to say so. The header-protection sample cannot read past the
buffer at any datagram floor: a CRYPTO frame header is three bytes before any
data, so the protected part always clears the nineteen the sample needs. ASan
and UBSan are clean at min_datagram=0. And the length varint is two bytes in
every reachable case, because a padded body stays under 16384 and an unpadded
one sizes its own field. Both are now explicit rather than implied, but neither
was fixing an overrun, and the first draft of these comments claimed otherwise.

Also: the OpenSSL calls in initial_packet() are checked, so a failure returns
nothing rather than a packet that is garbage after the header -- the target's
silence would have looked like a hold. And on_connect() in slow_quic now guards
the empty-hello underflow that on_timer() already guarded twenty lines below;
one of the two was wrong about the invariant and a reader could not tell which.

+++ ng/src/engine/engine.cpp
@@ -1069,7 +1069,11 @@ struct Engine::Impl {
     // Kernels round up and add bookkeeping overhead; Linux reports double the
     // request. Only flag the case where the granted buffer is so much larger
     // that the requested window is not meaningfully in force.
-    log.meta.window_overridden = granted > requested_rcvbuf_ * 4;
+    // Widened for the same reason classic needed it. ng caps -w/-y at 1048576
+    // so this cannot overflow today, but the bound and the arithmetic are in
+    // different files and nothing ties them together.
+    log.meta.window_overridden =
+        granted > static_cast<long long>(requested_rcvbuf_) * 4;
     if (!chatty()) return;
     interrupt_status();
     std::fprintf(stderr,
